+++ GoldenBack/GoldenBack.Tests/Factories/CustomWebApplicationFactory.cs
@@ -1,0 +1,47 @@
+// -----------------------------------------------------------------------
+// <copyright>
+//     Copyright 2024, Nikoden.IO All rights reserved.
+//     Author: Nicolas Denoël
+// </copyright>
+// -----------------------------------------------------------------------
+
+using GoldenBack.Infrastructure.DbContext;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace GoldenBack.Tests.Factories;
+
+public class CustomWebApplicationFactory<TProgram> : WebApplicationFactory<TProgram> where TProgram : class
+{
+    protected override void ConfigureWebHost(IWebHostBuilder builder)
+    {
+        builder.ConfigureServices(services =>
+        {
+            // Remove the existing ApplicationDbContext registration
+            var descriptor = services.SingleOrDefault(
+                d => d.ServiceType == typeof(DbContextOptions<ApplicationDbContext>));
+
+            if (descriptor != null)
+                services.Remove(descriptor);
+
+            // Add ApplicationDbContext with InMemory database for testing
+            services.AddDbContext<ApplicationDbContext>(options =>
+            {
+                options.UseInMemoryDatabase("factory-database");
+            });
+
+            var sp = services.BuildServiceProvider();
+
+            // Create a scope to obtain a reference to the database context
+            using (var scope = sp.CreateScope())
+            {
+                var scopedServices = scope.ServiceProvider;
+                var db = scopedServices.GetRequiredService<ApplicationDbContext>();
+
+                db.Database.EnsureCreated();
+            }
+        });
+    }
+}